// application/src/rendererApp/defaultSignGroup/types/dsg.ipcMessages.ts
enum dsgIpcMessages {
    UploadSigns = 'upload-signs',
    DSG_Create = 'dsg-create-message',
    DSG_Update = 'dsg-update-message',
    DSG_GetById = 'dsg-getById-message',
    DSG_DeleteById = 'dsg-deleteById-message',
    DSG_DeleteSignById = 'dsg-deleteSignById-message',
    DSG_GetList = 'dsg-getList-message',
    DSG_GetMenu = 'dsg-getMenu-message',
    DSG_AddSign = 'dsg-addSign-message',
    DSG_GetSign = 'dsg-getSign-message',
    DSG_EditSign = 'dsg-editSign-message',
    DSG_UploadImage = 'dsg-uploadImage-message',
    DSG_GetSignsList = 'dsg-getSignsList-message',
}
  
  export default dsgIpcMessages;